+++ lib/jkf/parser.rb
@@ -1,7 +1,5 @@
 module Jkf
-  # Define parser namespace
   module Parser
-    # Parse error
     class ParseError < StandardError; end
   end
 end